+++ src/views/Sections/index.js
@@ -1,4 +1,4 @@
-export { default as About } from "./About";
+export { default as Experience } from "./Experience";
 export { default as Contact } from "./Contact";
 export { default as Portfolio } from "./Portfolio";
 export { default as Services } from "./Services";
